requirements.txt
tensorflow~=2.3
Keras~=2.3
Pillow~=8.0.1
Flask~=1.1.2
numpy~=1.18
flask-cors~=3.0.9
gunicorn~=20.0.4